backend/requirements.txt
fastapi
pymongo
neo4j
uvicorn
python-dotenv
pydantic-settings
passlib[bcrypt] 
fastapi-login 
aiohttp
dnspython
python-multipart
nltk
pinecone
openai
boto3
bs4